Server-Side/requirements.txt
# Function dependencies, for example:
# package>=version
shapely 
numpy 
google-cloud-vision
google-cloud-storage
flask 
